changes due to manager splitting
