+++ BloodNight-core/src/main/java/de/eldoria/bloodnight/command/BloodNightCommand.java
@@ -2,8 +2,8 @@ package de.eldoria.bloodnight.command;
 
 import de.eldoria.bloodnight.command.bloodnight.*;
 import de.eldoria.bloodnight.config.Configuration;
-import de.eldoria.bloodnight.core.manager.MobManager;
-import de.eldoria.bloodnight.core.manager.NightManager;
+import de.eldoria.bloodnight.core.manager.mobmanager.MobManager;
+import de.eldoria.bloodnight.core.manager.nightmanager.NightManager;
 import de.eldoria.bloodnight.util.Permissions;
 import de.eldoria.eldoutilities.simplecommands.EldoCommand;
 import de.eldoria.eldoutilities.simplecommands.commands.DefaultDebug;
